feat: add confirmationationto delete todo item

+++ todo-list/src/components/TodoItem.js
@@ -10,7 +10,10 @@ const TodoItem = (props) => {
     }
 
     const handleDeleteItem = () => {
-        dispatch(deleteTodoItem(props.index))
+        const answer = window.confirm('Are you sure you wish to delete this item?');
+        if (answer) {
+            dispatch(deleteTodoItem(props.index))
+        }
     }
 
     return (
